move tests, bench and codegen deps to sub-package
This removes the codegen build dependency of the string_cache crate,
thereby minimizing dep for users that don't need codegen.

To allow the now external integration tests and benchmarks to test the
same things, public Atom::is_(static|dynamic|inline) -> bool methods
were also added.
