+++ Cargo.toml
@@ -7,7 +7,6 @@ description = "A string interning library for Rust, developed as part of the Ser
 license = "MIT / Apache-2.0"
 repository = "https://github.com/servo/string-cache"
 documentation = "https://docs.rs/string_cache/"
-build = "build.rs"
 
 # Do not `exclude` ./string-cache-codegen because we want to include
 # ./string-cache-codegen/shared.rs, and `include` is a pain to use
@@ -35,9 +34,8 @@ phf_shared = "0.7.4"
 new_debug_unreachable = "1.0"
 string_cache_shared = {path = "./shared", version = "0.3"}
 
-[dev-dependencies]
-rand = "0.4"
-string_cache_codegen = { version = "0.4", path = "./string-cache-codegen" }
-
-[build-dependencies]
-string_cache_codegen = { version = "0.4", path = "./string-cache-codegen" }
+[workspace]
+members = [
+    "string-cache-codegen",
+    "integration-tests",
+]
