+++ integration-tests/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "integration_tests"
+version = "0.0.1"
+authors = [ "The Servo Project Developers" ]
+build = "build.rs"
+publish = false
+
+[lib]
+doctest = false
+test = true
+
+[features]
+
+# Use unstable features to optimize space and time (memory and CPU usage).
+unstable = ["string_cache/unstable"]
+
+[dependencies]
+string_cache = { version = ">= 0.7.3, <= 9000", path = ".." }
+
+[dev-dependencies]
+rand = "0.4"
+string_cache_codegen = { version = "0.4", path = "../string-cache-codegen" }
+
+[build-dependencies]
+string_cache_codegen = { version = "0.4", path = "../string-cache-codegen" }
